Split PrunedLiveness::findBoundariesInBlock into fast paths
This allows dynamic dispatch into the fast paths. It's also easier to
understand the fast paths when written as separate loops.

This way, we can used MultiDefPrunedLiveness to compute ownership
liveness, but still benefit from the fact that the common case (owned
non-phi value) is much simpler and faster to compute.

This approach is necessary because we use stack-based data structures
within the liveness result, so we need to statically instantiate
something that can handle all cases.