Runtime: split out `extern "C"` marker from `SWIFT_RUNTIME_EXPORT`
Although the runtime exported symbols must be exported with "C" linkage
and decoration, we split out the externalisation in the C++ mode.  This
is required to ensure that we are able to attribute the functions
without following the more stringent rules that clang now enforces which
place the attributes following the identifier to which they appertain,
resulting in the attributes interleaving within the signature.  The
resulting spelling is more difficult to process for humans as well as
more difficult to identify as function declarations making code perusal
a challenge, particularly for the uninitiated.