ClangImporter: simplify logic for mutability attribute
Take the spellings for the attributes as parameters for the diagnostics
to allow us to reference the ignored the attribute properly.  Use direct
pointer checks instead of an unnecessary `llvm::Optional` around a
nullable pointer.  Use `llvm::StringRef` to simplify the comparison.
This avoids re-spelling the attribute parameters which have already been
checked by `isMutabilityAttr` for the contradiction.