[SILOpt] TempRValueElimination: 'let' copy source values cant be overwriten by apply instructions
TempRValueElimination checks if a copy of a value can be eliminated. Checking that the original
source value was not modified during the copy is one of the requirements for the elimination.
Prior to this change, TempRValueElimination checked if apply instructions could've
potentially modified the original source value, even if the source value was an immutable 'let'.
This prevented some optimization opportunities. With this change, we can recognize that 'let'
source value won't be modified by an apply, so that we can proceed with copy elimination.