add `_unsafeAssumeOnMainActor` as an adoption utility
This unofficial stdlib utility is designed to help you
adopt Swift concurrency progressively. You can use this
to help introduce a (totally) unsafe context which is allowed
to access things that are MainActor-isolated _without_ having
to `await` the MainActor operation closure.

This is totally unsafe in the sense that the runtime check
prior to invoking the MainActor operation is primarily there
for debugging purposes. You can customize its behavior to range
from ignoring failures, logging them, or triggering a fatal error.

So, even if you're not on the main thread, the operation may be
called anyway, which creates a race. I suggest not having this
utility appear in production releases of your software.

resolves rdar://101047674