[CXX-Interop] Improve support for anon enums with attributes
When cxx-interop is enabled, the NS_OPTIONS macro expands
to a typedef for a NSUInteger and an anonymous enum backed
by this new type. To achieve the intended semantics of NS_OPTION
when imported into Swift, the anonymous enum is imported under
the typedef's name, and references to the typedef are resolved to the enum.

Prior to this patch, imported name for the two entities
would not respect any attributes attached to the anonymous enum.
This patch fixes this and makes the way the two are associated less brittle.