Remove is_unique invariant we don't guarantee
The invariant is checking if the "cached" isUniquelyReferenced result is still valid.
The compiler can insert copies for lifetime adjustment during optimizations between
the time isUniquelyReferenced is first "cached" to the invariant, such copies don't escape
and don't really matter for COW. In such cases, this invariant will fire.

This PR turns off this invariant.

rdar://100530373