NFC: add code owners for C++ interop
This teaches GitHub to automatically add the three of us as reviewers to patches that alter the source files related to C++ interop.

This doesn't mean that all of the people added as reviewers must necessarily review a patch, is just makes it more straightforward for folks who don't have commit access to request reviews for their patches.