[move-only] Emit a "checker did not understand" error if copy propagation fails for some reason.
We already emit a "checker did not understand" error if our pattern matching at
the beginning failed. This just tightens the logic even more by including
another potential error condition.

NOTE: This was caught by the SILVerifier check that copy_value in canonical SIL
can not take a move only type as an argument. Its better to catch this with a
nice earlier diagnostic though.