Forcibly disable static constant array emission for now
This code currently uses the "empty array" metadata for static arrays (because
the "empty array" can never be freed and the metadata on the array storage is
not actually used in Swift).

But that doesn't work if the array is passed to Objective-C,
which does rely on the metadata.  So Objective-C code would
see bridged arrays of this type as empty.

Workaround for: rdar://101126543