Remove erroneous ELF note.
On ELF platforms we currently output an ELF note but with totally invalid
contents.  I tried experimenting to see whether it was possible to do it
properly and tie the result into ReflectionContext, but it turns out to be a
hard problem because of issues to do with cross-segment symbol references and
relocations.

The upshot is that I think for now it's best just to remove it.

We might revisit this in future at some point; it would be good to be able to
reliably find the Swift metadata just from the ELF Phdrs.

rdar://101749382