[AddressLowering] Handle try_apply results.
When a block argument is a terminator result from a try_apply, use the
ApplyRewriter to convert the try_apply.

In the case where the result is stored into an enum, with this change,
the init_enum_data_addr instruction is created prior to the try_apply
which is necessary in order for it to be passed as an argument to the
try_apply.