[AddressLowering] Take from cast destination.
When lowering an unconditional_checked_cast from an address-only value
to a loadable value, the value with which to replace the original is a
load of the destination address of the unconditional_checked_cast_addr
to which the instruction was lowered.  In the case of non-trivial
values, that load must be a take.